FastAPI/fastapi-2fa-tutorial/requirements.txt
fastapi
uvicorn
pyotp
python-dotenv